ci/skill-scanner/requirements.txt
openai>=1.30.0
pyyaml>=6.0.0
rich>=13.0.0
